+++ src/index.core.ts
@@ -20,8 +20,9 @@
  * @module
  */
 
-import { isIdentifier, isObject } from "@metreeca/core";
-import type { Indexed } from "./index.js";
+import { isBoolean, isIdentifier, isNumber, isObject, isString, isUnion } from "@metreeca/core";
+import { isIRI } from "@metreeca/core/resource";
+import type { Indexed, Literal, Reference } from "./index.js";
 
 
 /**
@@ -38,3 +39,24 @@ export function isIndexed<T>(value: unknown, is: (value: unknown) => value is T)
 	return isObject(value, (v, k) => isIdentifier(k) && is(v));
 }
 
+/**
+ * Checks if a value is a {@link Reference}.
+ *
+ * @param value The value to check
+ *
+ * @returns True if the value is an absolute IRI
+ */
+export function isReference(value: unknown): value is Reference {
+	return isIRI(value, "absolute");
+}
+
+/**
+ * Checks if a value is a {@link Literal}.
+ *
+ * @param value The value to check
+ *
+ * @returns True if the value is a boolean, finite number, or string
+ */
+export function isLiteral(value: unknown): value is Literal {
+	return isUnion(value, [isBoolean, isNumber, isString]);
+}